Rename `pact-jvm-mockk` to `pact-jvm-mock-mockk`, update dependencies in build files, and adjust imports in test contracts. Remove obsolete `pact-jvm-mockk` module files to streamline the project structure.
