+++ settings.gradle.kts
@@ -1,4 +1,4 @@
 rootProject.name = "pact-jvm-mock"
 include("pact-jvm-mock")
-include("pact-jvm-mockk")
+include("pact-jvm-mock-mockk")
 include("pact-jvm-mock-spring")
